feat: add solution with float for Task03 on operators

+++ src/main/java/com/rakovets/course/java/core/practice/operators/Task03.java
@@ -35,7 +35,7 @@ class Task03 {
      * @param depositAnnualPercentage ежегодный простой процент
      */
     static float getTotalDepositAmount(int depositAmount, int depositYears, int depositAnnualPercentage) {
-        double allProfits = depositAmount * depositAnnualPercentage / 100.f * depositYears;
-        return  (float) (allProfits + depositAmount);
+        float allProfits = depositAmount * depositAnnualPercentage / 100.f * depositYears;
+        return  allProfits + depositAmount;
     }
 }
